[#1] Adding the code to solve the exam
Adding the code to get data from https://api.opendata.esett.com/.
Adding the code to build an API with Swagger.
I am adding the code to get the Bank by ID.
Adding the code to save the data loaded from https://api.opendata.esett.com/ in the database and the necessary code for configuring EntityFramework with SQL database.
Adding the code with MVC architecture.
Adding CQRS and implementing a CRUD for the Bank entity.

Fix: #1